LocoAppBack/requirements.txt
fastapi==0.135.3
uvicorn[standard]==0.43.0
asyncpg==0.30.0
sqlalchemy[asyncio]==2.0.40
alembic==1.15.2
httpx==0.27.2
python-dotenv==1.0.1
pyyaml==6.0.2
